1. Implemented Categories form 2. Updated Products form(changed color and position for some labels) 3. Tables presenter - set table id after insert

+++ GlobalClassesStandard/Models/Families.cs
@@ -10,7 +10,7 @@ namespace GlobalClassesStandard.Models
     public class Families : ModelBase
     {
         [IgnoredProperty]
-        public Dictionary<int, string> LangToName { get; set; }
+        public Dictionary<long, string> LangToName { get; set; }
         public string Code{ get; set; }
     }
 }
